+++ test/UnitTests/UseCaseTests/Register/RegisterTests.cs
@@ -23,7 +23,7 @@ namespace UnitTests.UseCasesTests.Register
         [Fact]
         public void GivenNullInput_ThrowsException()
         {
-            var register = new Register(null, null, null, null, null, null);
+            var register = new Register(null, null, null, null, null, null, null);
             Assert.ThrowsAsync<Exception>(async() => await register.Execute(null));
         }
 
@@ -41,6 +41,7 @@ namespace UnitTests.UseCasesTests.Register
                 presenter,
                 _fixture.CustomerRepository,
                 _fixture.AccountRepository,
+                _fixture.UserRepository,
                 _fixture.UnitOfWork);
 
             await sut.Execute(new RegisterInput(
